major refactoring + EthereumTest uses Homestead config now
