+++ src/main/java/org/adridadou/ethereum/ethj/BlockchainConfig.java
@@ -1,4 +1,4 @@
-package org.adridadou.ethereum.blockchain;
+package org.adridadou.ethereum.ethj;
 
 import org.adridadou.ethereum.values.config.*;
