+++ src/main/java/org/adridadou/ethereum/ethj/provider/PrivateNetworkConfig.java
@@ -1,4 +1,4 @@
-package org.adridadou.ethereum.provider;
+package org.adridadou.ethereum.ethj.provider;
 
 import org.adridadou.ethereum.values.EthAccount;
 import org.adridadou.ethereum.values.EthValue;
