+++ src/main/java/org/adridadou/ethereum/rpc/EthereumRpcEventGenerator.java
@@ -1,4 +1,4 @@
-package org.adridadou.ethereum.blockchain;
+package org.adridadou.ethereum.rpc;
 
 import org.adridadou.ethereum.event.EthereumEventHandler;
 import org.ethereum.listener.EthereumListener;
